move gitlab authz config code to separate file
